pyproject.toml
[tool.poetry]
name = "dashpi"
version = "0.0.1"
description = "Todo"
authors = ["Farhad Ghamsari"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
rgbmatrixemulator = "^0.8.4"
ipython = "^8.13.2"
pillow = "^9.5.0"
python-dateutil = "^2.8.2"
requests = "^2.30.0"
numpy = "^1.24.3"
pyowm = "^3.3.0"
websocket-client = "^1.5.1"
gpiozero = "^1.6.2"
scipy = "^1.10.1"
spotipy = "^2.23.0"
bdfparser = "^2.2.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
